+++ includes/class-cocart-session-handler.php
@@ -121,7 +121,17 @@ class CoCart_Session_Handler extends WC_Session_Handler {
 			$cart_key = (string) trim( sanitize_key( wp_unslash( $_SERVER['HTTP_COCART_API_CART_KEY'] ) ) ); // phpcs:ignore WordPress.Security.NonceVerification.Recommended
 		}
 
-		return $cart_key;
+		/**
+		 * Filter allows the cart key to be overridden.
+		 *
+		 * Developer Note: Really only here so I don't have to create
+		 * a new session handler to inject a customer ID with the POS Support Add-on.
+		 *
+		 * @since 4.x.x Introduced.
+		 *
+		 * @ignore Function ignored when parsed into Code Reference.
+		 */
+		return apply_filters( 'cocart_requested_cart_key', $cart_key );
 	} // END get_requested_cart()
 
 	/**
@@ -153,7 +163,7 @@ class CoCart_Session_Handler extends WC_Session_Handler {
 			$this->_data = $this->get_session_data();
 
 			// If the user logs in, update session.
-			if ( is_user_logged_in() && $current_user_id !== $this->_customer_id ) {
+			if ( is_user_logged_in() && $this->is_user_customer( $current_user_id ) && $current_user_id !== $this->_customer_id ) {
 				$guest_session_id   = $this->_customer_id;
 				$this->_customer_id = $current_user_id;
 				$this->save_data( $guest_session_id );
@@ -171,6 +181,33 @@ class CoCart_Session_Handler extends WC_Session_Handler {
 			$this->_data        = $this->get_session_data();
 		}
 	} // END init_session_cocart()
+
+	/**
+	 * Detect if the user is a customer.
+	 *
+	 * @since 4.x.x Introduced.
+	 *
+	 * @param int $user_id The user ID.
+	 *
+	 * @return bool Returns true if user is a customer, otherwise false.
+	 */
+	public function is_user_customer( $user_id ) {
+		if ( ! is_int( $user_id ) || 0 === $user_id ) {
+			return false;
+		}
+
+		$current_user = get_userdata( $user_id );
+
+		if ( ! empty( $current_user ) ) {
+			$user_roles = $current_user->roles;
+
+			if ( in_array( 'customer', $user_roles, true ) ) {
+				return true;
+			}
+		}
+
+		return false;
+	} // END is_user_customer()
 
 	/**
 	 * Return true if the current customer has an active cart.